% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PropensityModeling.R
\name{calculateIPTW}
\alias{calculateIPTW}
\title{Calculate Inverse Probability of Treatment Weights (IPTW)}
\usage{
calculateIPTW(propensityScores, treatment, stabilized = TRUE)
}
\arguments{
\item{propensityScores}{Vector of propensity scores}

\item{treatment}{Binary treatment indicator}

\item{stabilized}{Logical; use stabilized weights (default: TRUE)}
}
\value{
Vector of IPTW weights
}
\description{
Calculate Inverse Probability of Treatment Weights (IPTW)
}
