% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PropensityModeling.R
\name{calculateOverlapWeights}
\alias{calculateOverlapWeights}
\title{Calculate Overlap Weights}
\usage{
calculateOverlapWeights(propensityScores, treatment)
}
\arguments{
\item{propensityScores}{Vector of propensity scores}

\item{treatment}{Binary treatment indicator}
}
\value{
Vector of overlap weights
}
\description{
Calculates overlap weights (Li et al. 2018) which optimize balance
and precision by focusing on the population with equipoise.
}
