% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diagnostics.R
\name{calculateSMD}
\alias{calculateSMD}
\title{Calculate Standardized Mean Difference (SMD)}
\usage{
calculateSMD(covariates, treatment, weights = NULL)
}
\arguments{
\item{covariates}{Matrix of covariates}

\item{treatment}{Binary treatment indicator}

\item{weights}{Optional weights}
}
\value{
Vector of SMDs
}
\description{
Calculate Standardized Mean Difference (SMD)
}
\keyword{internal}
