% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConnectionHandler.R
\name{createSyTrialConnection}
\alias{createSyTrialConnection}
\title{Create Database Connection for SyTrial Analysis}
\usage{
createSyTrialConnection(
  connectionDetails,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortTable = "cohort",
  tempEmulationSchema = NULL
)
}
\arguments{
\item{connectionDetails}{DatabaseConnector connection details object}

\item{cdmDatabaseSchema}{Schema name containing OMOP CDM tables}

\item{cohortDatabaseSchema}{Schema name for cohort tables}

\item{cohortTable}{Name of the cohort table}

\item{tempEmulationSchema}{Schema for temporary tables (Oracle/BigQuery)}
}
\value{
A SyTrialConnection object
}
\description{
Establishes connection to OMOP CDM database and validates schema structure.
}
\examples{
\dontrun{
connectionDetails <- DatabaseConnector::createConnectionDetails(
  dbms = "postgresql",
  server = "localhost/ohdsi",
  user = "postgres",
  password = "password"
)

syTrialConn <- createSyTrialConnection(
  connectionDetails = connectionDetails,
  cdmDatabaseSchema = "cdm_synpuf",
  cohortDatabaseSchema = "results",
  cohortTable = "cohort"
)
}
}
