% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConnectionHandler.R
\name{disconnectSyTrial}
\alias{disconnectSyTrial}
\title{Disconnect SyTrial Connection}
\usage{
disconnectSyTrial(syTrialConnection)
}
\arguments{
\item{syTrialConnection}{A SyTrialConnection object}
}
\description{
Disconnect SyTrial Connection
}
