% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diagnostics.R
\name{generateDiagnosticReport}
\alias{generateDiagnosticReport}
\title{Generate Comprehensive Diagnostic Report}
\usage{
generateDiagnosticReport(syTrialResult, outputDir = ".")
}
\arguments{
\item{syTrialResult}{SyTrial analysis result object}

\item{outputDir}{Directory to save diagnostic outputs}
}
\value{
List of diagnostic objects
}
\description{
Generate Comprehensive Diagnostic Report
}
