% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SyTrialAnalysis.R
\name{print.SyTrialResult}
\alias{print.SyTrialResult}
\title{Print SyTrial Result}
\usage{
\method{print}{SyTrialResult}(x, ...)
}
\arguments{
\item{x}{SyTrialResult object}

\item{...}{Additional arguments}
}
\description{
Print SyTrial Result
}
