% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SyTrialAnalysis.R
\name{summary.SyTrialResult}
\alias{summary.SyTrialResult}
\title{Summary of SyTrial Result}
\usage{
\method{summary}{SyTrialResult}(object, ...)
}
\arguments{
\item{object}{SyTrialResult object}

\item{...}{Additional arguments}
}
\description{
Summary of SyTrial Result
}
