% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diagnostics.R
\name{createKaplanMeierPlot}
\alias{createKaplanMeierPlot}
\title{Create Kaplan-Meier Survival Curves}
\usage{
createKaplanMeierPlot(
  data,
  time,
  event,
  treatment,
  weights = NULL,
  title = "Kaplan-Meier Survival Curves",
  fileName = NULL
)
}
\arguments{
\item{data}{Data frame with survival data}

\item{time}{Character; name of time variable}

\item{event}{Character; name of event indicator variable}

\item{treatment}{Character; name of treatment variable}

\item{weights}{Optional weights for weighted survival curves}

\item{title}{Plot title}

\item{fileName}{Optional file name to save plot}
}
\value{
A survminer ggsurvplot object
}
\description{
Generates Kaplan-Meier curves comparing treatment and synthetic control arms.
}
\examples{
## Not run: 
## library(survival)
## library(survminer)
## set.seed(1)
## n <- 50
## df <- data.frame(
##   time = rexp(n, 0.1),
##   event = rbinom(n, 1, 0.7),
##   trt = sample(c("Treatment", "SyntheticControl"), n, replace = TRUE)
## )
## createKaplanMeierPlot(
##   data = df,
##   time = "time",
##   event = "event",
##   treatment = "trt",
##   title = "Kaplan-Meier Survival Curves"
## )
## End(Not run)
}
